// src/domain/uhTipo/types.ts
export type UHTipo = {
  iduhtipo: number;
  cduhtipo: string;
  nmuhtipo: string;
  dsuhtipo: string;
  iduhtipogrupo: number;
  qtleito: number;
  dhinclusao: string;
  dhalteracao: string;
  fltipocobranca: number;
  iduhtipogrupo_1: number;
  nmtipogrupo: string;
  flsituacao: number;
  idempresa: number;
  idpai: number;
  empresa_dsabreviatura: string;
};